Move results to `src/calculate`.
